Add methods to add losers to bracket while in-progress; basic test for bracket
This commit adds a public method to the losers bracket class to add losing
teams from the winners bracket to the appropriate place in the losers bracket,
as determined by a list of positions generated at object construction by a
new private method. A stub test module for double-elimination cups has been
added with a simple test regarding the number of rounds in a losers bracket
given the number of rounds in the winners bracket.