Implement double-elimination cups for powers of two
This change introduces double-elimination cups for powers of two.
The losers bracket, introduced in an earlier commit, has a fixed format.
An option is given to replay the final match if the winning team came
from the losers bracket. This is enabled by default.

Tests have been added, mostly copied from the single-elimination cup.
The dependency on competitions-match has been bumped to v0.1.3 to allow
for the use of the loser field included by default in that version.