New "standard" single-elimination cup class
StandardSingleEliminationCup allows for byes in the first round,
enabling single-elimination cup brackets for any number of teams.

While this new class technically obsoletes the original
PowerOfTwoSingleEliminationCup, that class will be kept in the code
for now.

This first beta begins the 0.4 release series. New, more relevant
tests for the added code will be written before the final release.