Code refactoring and updated interface
The setuptools entry point registration has been removed. Classes should
now be used directly. The classes and tests were updated to reflect this
change. Matches are also used directly as of competitions-match 0.3. As
such, the match class must now be passed to the cup constructor.

Also, common cup constructor code was refactored to the parent constructor.

This serves as the first (and hopefully only) beta release for version 0.5.
No new features were added, but I figured the interface change was enough
to necessitate the version bump.