Rename and merge cup modules
The single-elimination cups and the double-elimination brackets were
merged into a single module each. All remaining cup modules were renamed
to more Pythonic module names. The tests were updated to reflect the new
module names.

This will be released as part of the 0.5 series to avoid another version
bump. This will be version 0.5.1.